dom/workers/test/test_fileSlice.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=664783
-->
<window title="Mozilla Bug 664783"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="dom_worker_helper.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=664783"
     target="_blank">Mozilla Bug 664783</a>

  <div id="content" style="display: none">
    <input id="fileList" type="file"></input>
  </div>

  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 664783 **/

  var fileNum = 0;

  /**
   * Create a file which contains the given data and optionally adds the specified file extension.
   */
  function createFileWithData(fileData, /** optional */ extension) {
    var testFile = Components.classes["@mozilla.org/file/directory_service;1"]
                       .getService(Components.interfaces.nsIProperties)
                       .get("ProfD", Components.interfaces.nsIFile);
    var fileExtension = (extension == undefined) ? "" : "." + extension;
    testFile.append("workerSlice" + fileNum++ + fileExtension);

    var outStream = Components.classes["@mozilla.org/network/file-output-stream;1"]
                        .createInstance(Components.interfaces.nsIFileOutputStream);
    outStream.init(testFile, 0x02 | 0x08 | 0x20, // write, create, truncate
                   0666, 0);
    outStream.write(fileData, fileData.length);
    outStream.close();

    var fileList = document.getElementById('fileList');
    fileList.value = testFile.path;

    return fileList.files[0];
  }

  /**
   * Starts a worker which slices the blob to the given start offset and optional end offset and
   * content type. It then verifies that the size and type of the sliced blob is correct.
   */
  function createSlice(blob, start, expectedLength, /** optional */ end, /** optional */ contentType) {
    var worker = new Worker("fileSlice_worker.js");

    worker.onerror = function(event) {
      ok(false, "Worker had an error: " + event.data);
      finish();
    };

    worker.onmessage = function(event) {
      is(event.data.size, expectedLength, "size property of slice is incorrect.");
      is(event.data.type, contentType ? contentType : blob.type, "type property of slice is incorrect.");
      finish();
    };

    var params = {blob: blob, start: start, end: end, contentType: contentType};
    worker.postMessage(params);
    waitForWorkerFinish();
  }

  // Empty file.
  createSlice(createFileWithData(""), 0, 0, 0);

  // Typical use case.
  createSlice(createFileWithData("Hello"), 1, 1, 2);

  // Longish file.
  var text = "";
  for (var i = 0; i < 10000; ++i) {
    text += "long";
  }
  createSlice(createFileWithData(text), 2000, 2000, 4000);

  // Slice to different type.
  createSlice(createFileWithData("text", "txt"), 0, 2, 2, "image/png");

  // Length longer than blob.
  createSlice(createFileWithData("text"), 0, 4, 20);

  // Start longer than blob.
  createSlice(createFileWithData("text"), 20, 0, 4);

  // No optional arguments
  createSlice(createFileWithData("text"), 0, 4);
  createSlice(createFileWithData("text"), 2, 2);

  ]]>
  </script>
</window>
